增加了多媒体操作示例，打开相机，读取图片，发送Notification ，音频视频还未实现

+++ app/src/main/java/com/example/leo/activitystudy/TelephoneBook/TelephoneBookAdapter.java
@@ -27,7 +27,7 @@ public class TelephoneBookAdapter extends ArrayAdapter<TelephoneBook> {
         View view;
         RecyclerView.ViewHolder viewHolder;
         if(convertView==null) {
-            view = LayoutInflater.from(getContext()).inflate(resourceImgId, parent, false);
+            view = LayoutInflater.from(getContext()).inflate((int) resourceImgId, parent, false);
          //   viewHolder = new RecyclerView.ViewHolder();
         }
         else{
